Moved autocomplete to CLI_Controller
