+++ src/Controllers/cli_controller.py
@@ -2,21 +2,58 @@ from Models.uml_diagram import UML_Diagram
 from Models.error_handler import Error_Handler
 from Models.uml_parser import parse
 from Views.cli_view import CLI_View
+from prompt_toolkit import prompt
+from prompt_toolkit.completion import NestedCompleter
 
 class CLI_Controller:
 
     def __init__(self, diagram:UML_Diagram = UML_Diagram(), saved:bool = False):
         self._diagram = diagram
-        self._tab_complete = CLI_View()
         self._should_quit = False
+        self._completer = self.setup_autocomplete()  
     
-    def update(self, input:str):
+    def update(self):
+        input = prompt("Command: ", completer=self._completer).strip().lower()
         if len(input.strip()) > 0:
             data = parse(self._diagram, input)
             if not hasattr(CLI_View, str(input[:4])):
                 return data[0](*data[1:])
             #print help menu
             print(data[0](*data[1:]))
+    
+    def setup_autocomplete(self):
+        """
+        Sets up autocomplete functionality using the NestedCompleter.
+        """
+        return NestedCompleter.from_nested_dict({
+        'add': {
+            'class': None,
+            'field': None,
+            'method': None,
+            'relationship': None
+        },
+        'delete': {
+            'class': None,
+            'field': None,
+            'method': None,
+            'relationship': None
+        },
+        'rename':{
+            'class': None,
+            'field': None,
+            'method': None
+        },
+        'list':{
+            'class': None,
+            'classes': None,
+            'relationship': None,
+            'relationships': None
+        },
+        'save': None,
+        'load': None,
+        'exit': None
+        })  
+
